+++ packages/cheeket/lib/bind-strategy/bind-strategy.ts
@@ -1,5 +1,5 @@
-import Context from "./context";
-import Next from "./next";
+import Context from "../context";
+import Next from "../next";
 
 interface BindStrategy<T, U> {
   bind(context: Context<T>, response: U): Promise<void> | void;
