+++ packages/cheeket/lib/bind-strategy/index.ts
@@ -1,0 +1,3 @@
+export { default as BindStrategy } from "./bind-strategy";
+export { default as bindArray } from "./bind-array";
+export { default as bindObject } from "./bind-object";
