+++ packages/cheeket/lib/index.ts
@@ -1,4 +1,3 @@
-export { default as Middleware } from "./middleware";
 export { default as Resolver } from "./resolver";
 export { default as Context } from "./context";
 export { default as Register } from "./register";
@@ -8,18 +7,13 @@ export { default as Type } from "./type";
 export { default as Abstract } from "./abstract";
 export { default as Next } from "./next";
 export { default as Done } from "./done";
-
 export { default as Factory } from "./factory";
-export { default as BindStrategy } from "./bind-strategy";
-
-export { default as bindArray } from "./bind-array";
-export { default as bindObject } from "./bind-object";
 
 export { default as InternalTokens } from "./internal-tokens";
 export { default as InternalEvents } from "./internal-events";
 
 export { default as AsyncEventEmitter } from "./async-event-emitter";
 
-export { default as inContainerScope, InContainerScope } from "./in-container-scope";
-export { default as inGlobalScope, InGlobalScope } from "./in-global-scope";
-export { default as inRequestScope, InRequestScope } from "./in-request-scope";
+export * from "./bind-strategy";
+export * from "./scope";
+export { Middleware } from "./middleware";
