+++ packages/cheeket/lib/middleware/middleware.ts
@@ -1,5 +1,5 @@
-import Context from "./context";
-import Next from "./next";
+import Context from "../context";
+import Next from "../next";
 
 type Middleware<T> = (context: Context<T>, next: Next) => Promise<void> | void;
