+++ packages/cheeket/lib/middleware/proxy.ts
@@ -1,5 +1,5 @@
-import MiddlewareStorage from "../middleware-storage";
-import Middleware from "../middleware";
+import MiddlewareStorage from "./middleware-storage";
+import Middleware from "./middleware";
 import Token from "../token";
 
 function proxy<T>(storage: MiddlewareStorage, token: Token<T>): Middleware<T> {
