+++ packages/cheeket/lib/scope/in-global-scope.ts
@@ -1,9 +1,9 @@
-import Middleware from "./middleware";
-import InternalTokens from "./internal-tokens";
-import InternalEvents from "./internal-events";
-import Factory from "./factory";
-import BindStrategy from "./bind-strategy";
-import AsyncLock from "./async-lock";
+import Middleware from "../middleware/middleware";
+import InternalTokens from "../internal-tokens";
+import InternalEvents from "../internal-events";
+import Factory from "../factory";
+import BindStrategy from "../bind-strategy/bind-strategy";
+import AsyncLock from "../async-lock";
 
 type InGlobalScope<T> = Middleware<T>;
