+++ packages/cheeket/lib/scope/in-request-scope.ts
@@ -1,8 +1,8 @@
-import Middleware from "./middleware";
-import InternalTokens from "./internal-tokens";
-import InternalEvents from "./internal-events";
-import Factory from "./factory";
-import BindStrategy from "./bind-strategy";
+import Middleware from "../middleware/middleware";
+import InternalTokens from "../internal-tokens";
+import InternalEvents from "../internal-events";
+import Factory from "../factory";
+import BindStrategy from "../bind-strategy/bind-strategy";
 
 type InRequestScope<T> = Middleware<T>;
